Stylus support, numbers on view & more
So... This have been mainly done at the end of my holidays and during train
travels, so it's clearly untested and may contain bugs.

I've added two smaller picrosses to debug the solver and the view.

The view now has the numbers on lines and columns and accept stylus input using the jpen
library. It can detect if the picross have been completed.

TODO:
 - Implement the other methods of solving (corner method...)
 - Add a standard input (mouse/keyboard)
 - Double the lines each five columns/rows
 - Add a menu to load, save ...
 - Get a better way to draw the numbers