Finally no more OutOfMemoryException.
So, I developed a new method that doesn't require that much memory.
So... It's cool, but still need improvement, as it is still slow for the
beach_side.txt. However it's way faster.

Still need a better way than trying every possibility.
I need to take out the solving part from Row.java too.
(yeah, I know, should not be here in the first place)